+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Pass `retention_days` in the Kafka profile messages. ([#3362](https://github.com/getsentry/relay/pull/3362))
 - Support and expose namespaces for metric rate limit propagation via the `x-sentry-rate-limits` header. ([#3347](https://github.com/getsentry/relay/pull/3347))
 - Tag span duration metric by group for all ops supporting description scrubbing. ([#3370](https://github.com/getsentry/relay/pull/3370))
+- Copy transaction tags to segment. ([#3386](https://github.com/getsentry/relay/pull/3386))
 - Route spans according to trace_id. ([#3387](https://github.com/getsentry/relay/pull/3387))
 
 ## 24.3.0